requirements.txt
langchain==0.3.24
langchain-core==0.3.56
langchain-openai==0.3.15
langchain-chroma==0.2.3
langchain-community==0.3.23
chromadb==0.6.3
tiktoken==0.9.0
openai==1.76.2
PyYAML==6.0.2
pypdf==5.4.0
tqdm==4.67.1
markdown-it-py==3.0.0
mdurl==0.1.2
httpx==0.28.1
httpx-sse==0.4.0
pillow==11.2.1
PyMuPDF==1.23.20
streamlit==1.45.0
dataclasses-json==0.6.7
typing_extensions==4.13.2
sentence-transformers
